pylbo: refactored spectra visualisations
